segmentation/requirements.txt
numpy
torch
torchvision
opencv-python
pillow
matplotlib
pathlib
scikit-learn
scikit-image
einops
pydensecrf
